requirements.txt
python-socketio
python-dotenv
numpy
opencv-python
requests